tests/test_bot.py
import pytest
import pytest_asyncio
from src.bot import FinanceBot
from src.models import TransactionType
from unittest.mock import AsyncMock, MagicMock, patch
from telegram import Update, User, Message, Chat
from telegram.ext import ContextTypes


@pytest.fixture
def bot():
    return FinanceBot()


@pytest.fixture
def mock_update():
    update = MagicMock(spec=Update)
    update.effective_user = MagicMock(spec=User)
    update.effective_user.id = 12345
    update.effective_user.username = "test_user"
    update.message = MagicMock(spec=Message)
    update.message.chat = MagicMock(spec=Chat)
    update.message.reply_text = AsyncMock()
    return update


@pytest.fixture
def mock_context():
    return MagicMock(spec=ContextTypes.DEFAULT_TYPE)


def parse_amount(amount_str: str) -> float:
    """Парсинг суммы с учетом пробелов"""
    # Удаляем все нецифровые символы, кроме точки, запятой и пробелов
    clean_str = "".join(c for c in amount_str if c.isdigit() or c in "., ")
    # Удаляем все пробелы
    clean_str = clean_str.replace(" ", "")
    # Заменяем запятую на точку
    clean_str = clean_str.replace(",", ".")
    return float(clean_str)


def split_message(message: str) -> tuple[str, str]:
    """Разделяет сообщение на сумму и описание"""
    # Пропускаем знак +/- в начале
    message = message[1:].strip()

    # Ищем первое слово, которое не является частью суммы
    parts = message.split()
    amount_parts = []
    description_parts = []

    for part in parts:
        # Если часть содержит цифры или является обозначением валюты
        if any(c.isdigit() for c in part) or part.lower() in [
            "руб",
            "руб.",
            "р.",
            "рублей",
        ]:
            amount_parts.append(part)
        else:
            description_parts.append(part)
            break

    # Добавляем оставшиеся части к описанию
    description_parts.extend(parts[len(amount_parts) + len(description_parts) :])

    return " ".join(amount_parts), " ".join(description_parts)


def test_parse_expense_message():
    """Тест парсинга сообщений с расходами"""
    test_cases = [
        ("-500 продукты", (500, "продукты")),
        ("-1000.50 такси", (1000.50, "такси")),
        ("-1 234 567 аренда квартиры", (1234567, "аренда квартиры")),
        ("-100руб. кофе", (100, "кофе")),
        ("-50р. проезд", (50, "проезд")),
        ("-42.50 рублей обед", (42.50, "обед")),
        ("-15 320 на счетах", (15320, "на счетах")),
        ("-357 за такси", (357, "за такси")),
        ("-400 за симку", (400, "за симку")),
        ("-48 на дорогу", (48, "на дорогу")),
        ("-302 вкусно и точка", (302, "вкусно и точка")),
        ("-500 на впн", (500, "на впн")),
    ]

    for message, expected in test_cases:
        amount_str, description = split_message(message)
        amount = parse_amount(amount_str)
        assert (
            amount,
            description,
        ) == expected, f"Ошибка парсинга: '{message}' -> получено ({amount}, {description}), ожидалось {expected}"


def test_parse_income_message():
    """Тест парсинга сообщений с доходами"""
    test_cases = [
        ("+5000 зарплата", (5000, "зарплата")),
        ("+1000.50 фриланс", (1000.50, "фриланс")),
        ("+1 234 567 продажа", (1234567, "продажа")),
        ("+100руб. возврат", (100, "возврат")),
        ("+50р. кэшбэк", (50, "кэшбэк")),
        ("+42.50 рублей проценты", (42.50, "проценты")),
        ("+15 320 на счетах банков", (15320, "на счетах банков")),
        ("+300 на впн", (300, "на впн")),
        ("+100 за кофе", (100, "за кофе")),
    ]

    for message, expected in test_cases:
        amount_str, description = split_message(message)
        amount = parse_amount(amount_str)
        assert (
            amount,
            description,
        ) == expected, f"Ошибка парсинга: '{message}' -> получено ({amount}, {description}), ожидалось {expected}"


def test_category_detection():
    """Тест определения категорий по ключевым словам"""
    bot = FinanceBot()
    test_cases = [
        ("продукты в магазине", "Продукты"),
        ("такси до работы", "Транспорт"),
        ("аренда квартиры", "Жилье"),
        ("кино с друзьями", "Развлечения"),
        ("лекарства в аптеке", "Здоровье"),
        ("новая куртка", "Одежда"),
        ("курсы python", "Образование"),
        ("новый телефон", "Техника"),
        ("подарок маме", "Подарки"),
        ("vpn сервис", "Связь"),
        ("вкусно и точка", "Продукты"),
        ("ростикс", "Продукты"),
        ("мартирос обед", "Продукты"),
        ("на дорогу до Мытищ", "Транспорт"),
        ("за симку", "Связь"),
        ("на впн", "Связь"),
        ("шоколадки на 8 марта", "Подарки"),
    ]

    for description, expected_category in test_cases:
        found_category = "Другое"
        # Сначала проверяем специальные случаи
        if "8 марта" in description.lower():
            found_category = "Подарки"
        else:
            # Затем проверяем обычные ключевые слова
            for category, keywords in bot.category_keywords.items():
                if any(keyword.lower() in description.lower() for keyword in keywords):
                    found_category = category
                    break
        assert (
            found_category == expected_category
        ), f"Для описания '{description}' ожидалась категория '{expected_category}', получена '{found_category}'"


@pytest.mark.asyncio
async def test_process_transaction_message(bot, mock_update, mock_context):
    """Тест обработки сообщений о транзакциях"""
    # Тестируем расход
    mock_update.message.text = "-500 такси"
    await bot.process_transaction_message(mock_update, mock_context)
    mock_update.message.reply_text.assert_called_once()
    mock_update.message.reply_text.reset_mock()

    # Тестируем доход
    mock_update.message.text = "+1000 зарплата"
    await bot.process_transaction_message(mock_update, mock_context)
    mock_update.message.reply_text.assert_called_once()
    mock_update.message.reply_text.reset_mock()

    # Тестируем некорректный формат
    mock_update.message.text = "это не транзакция"
    await bot.process_transaction_message(mock_update, mock_context)
    mock_update.message.reply_text.assert_not_called()


@pytest.mark.asyncio
async def test_add_transaction_start(bot, mock_update, mock_context):
    """Тест начала диалога добавления транзакции"""
    result = await bot.add_transaction_start(mock_update, mock_context)
    assert result == bot.CHOOSING_TYPE
    mock_update.message.reply_text.assert_called_once()
